resume_platform/requirements.txt
anthropic
openai
pdfplumber
python-docx
rich
streamlit
python-dotenv
typer
plotly
pydantic
pytesseract
Pillow
pdf2image